Move block proposal code into its own module
